requirements.txt
pyspark==3.5.1
python-dotenv==1.0.1
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-auth==2.29.0
requests==2.31.0
